Add taus to the model state dict for EnsembleMapSaasSingleTaskGP
Summary: `EnsembleMapSaasSingleTaskGP` doesn't save the `taus` as part of the state dict. This is confusing as creating a new model followed by loading the state dict may result in the learned model hyperparameters being suboptimal and thus changing if calling `fit_gpytorch_mll` again. This diff adds `taus` as a buffer to avoid this issue.

Differential Revision: D85768264